plotly check

+++ streamlit_app.py
@@ -463,7 +463,7 @@ if col4.button('Train model!') and st.session_state.legit and col_to_drop.count(
         fig.savefig(buf, format="png")
         st.image(buf)
 
-        middle = (max(pereds) - min(pereds)) / 10
+        middle = int((max(pereds) - min(pereds)) / 10)
         fig = ff.create_distplot([pereds], ['Regressor predictions'], show_curve=False, colors=['red'],
                                  bin_size=list(range(min(pereds), max(pereds), middle)))
